Reorganize into utils header and source files, clean up main code
